requirements.txt
Flask==3.0.3
Flask_Cors==4.0.0
fuzzywuzzy==0.18.0
pandas==2.0.3
python_dateutil==2.8.2
Requests==2.31.0
scikit_learn==1.4.2
torch==2.1.0
transformers==4.36.2
python-Levenshtein==0.25.1